pyproject.toml
[project]
name = "potato-registry"
version = "0.1.0"
description = "A registry for managing python packages."
authors = [{ name = "Léo Brossard", email = "leo_brossard@hotmail.fr" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "jwt>=1.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "standard>=1.0.3",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = ["pytest>=9.0.1", "pytest-asyncio>=1.3.0"]
